Removed StudentDisabilityProfile as a separate component
